requirements.txt
pysnc==1.0.5
python_dateutil==2.8.2
requests==2.27.1
